requirements.txt
llama-cpp-python==0.2.90
docker==7.1.0
tqdm==4.67.1
numpy==1.26.4